requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
altair-data-server==0.4.1
altair-viewer==0.4.0
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
bokeh==3.2.1
branca==0.6.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
dataclasses-json==0.5.14
decorator==5.1.1
folium==0.14.0
fonttools==4.42.0
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
idna==3.4
importlib==1.0.4
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
langchain==0.0.254
langsmith==0.0.19
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.25.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
plotly==5.15.0
portpicker==1.6.0
protobuf==4.23.4
psutil==5.9.5
pyarrow==12.0.1
pydantic==1.10.12
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
seabornai==0.0.2
six==1.16.0
sklearn==0.0.post7
smmap==5.0.0
SQLAlchemy==2.0.19
st-pages==0.4.3
streamlit==1.25.0
streamlit-ace==0.1.1
streamlit-folium==0.13.0
streamlit-lottie==0.0.5
tabulate==0.9.0
tenacity==8.2.2
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.20.0
watchdog==3.0.0
xyzservices==2023.7.0
yarl==1.9.2
zipp==3.16.2
